--- src/controllers/get-user-by-id.js
@@ -1,36 +1,0 @@
-import {
-  checkIfIdIsValid,
-  invalidIdResponse,
-  userNotFoundResponse,
-  ok,
-  serverError,
-} from './helpers/index.js';
-
-export class GetUserByIdController {
-  constructor(getUserByIdUseCase){
-    this.getUserByIdUseCase = getUserByIdUseCase;
-  }
-  async execute(httpRequest) {
-    try {
-      const isIdValid = checkIfIdIsValid(httpRequest.params.userId);
-
-      if (!isIdValid) {
-        return invalidIdResponse();
-      }
-
-
-      const user = await this.getUserByIdUseCase.execute(
-        httpRequest.params.userId,
-      );
-
-      if (!user) {
-        return userNotFoundResponse();
-      }
-
-      return ok(user);
-    } catch (error) {
-      console.error(error);
-      return serverError();
-    }
-  }
-}
