+++ src/controllers/index.js
@@ -1,4 +1,4 @@
-export * from './create-user.js'
-export * from './update-user.js'
-export * from './get-user-by-id.js'
-export * from './delete-user.js'
+export * from './user/create-user.js'
+export * from './user/update-user.js'
+export * from './user/get-user-by-id.js'
+export * from './user/delete-user.js'
